Vancomycin Model-informed precision dosing (MIPD)
Model-informed precision dosing (MIPD) of vancomycin integrating Bayesian estimation is important to ensure adequate vancomycin exposure, especially in children. Coupled to the lack of pharmacy- regulated MIPD programs, the availability of Bayesian software is limited due to cost restraint in a low-resource country. The project To compares the accuracy and precision of vancomycin Bayesian AUC24 estimations.